requirements.txt
Django==1.6.5
httplib2==0.8
restclient==0.11.0
feedparser==5.1.3
Markdown==2.3.1
simplejson==3.3.0
smartypants==1.6.0.3
uuid==1.30
psycopg2==2.5.1
Pillow==2.3.0
nose==1.3.0
versiontools==1.9.1
statsd==2.0.2
pep8==1.4.6
pyflakes==0.7.3
mccabe==0.2.1
flake8==2.0
tornado==3.1
BeautifulSoup==3.2.1
cssselect==0.8
lxml==3.2.3
fuzzywuzzy==0.2
sure==1.2.2
ipython==0.13.2
ipdb==0.7
selenium==2.31.0
extras==0.0.3
python-mimeparse==0.1.4
testtools==0.9.35
testscenarios==0.4
python-subunit==0.0.18
lettuce==0.2.19ccnmtl
rdflib==3.1.0
coverage==3.5.3
requests==0.13.3
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0
six==1.3.0
factory_boy==2.1.1
python-ldap==2.4.13
sqlparse==0.1.11

djangowind==0.12.1
sorl==3.1
tagging==0.3-pre
typogrify==2.0.0
django-indexer==0.3.0
django-templatetag-sugar==0.1
django-paging==0.2.5
South==0.8.4
django-annoying==0.7.6
django-nose==1.2
django-appconf==0.6
django_compressor==1.3
django-statsd-mozilla==0.3.11
raven==4.0.3
django-bootstrap-form==2.0.3
django-debug-toolbar==1.1
django-waffle==0.9.1
django-jenkins==0.15.0
django-smoketest==0.2.2
django-extensions==1.2.5
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==0.8.0
django-registration==1.0
django-treebeard==2.0b2
django-pagetree==0.14.5
django-pageblocks==0.5.11
django-quizblock==0.8.1
django-markwhat==1.0
